fix web build + move function to builder to library
