+++ crates/bevy_plugin/src/dialogue_runner/builder.rs
@@ -4,7 +4,6 @@ use crate::line_provider::SharedTextProvider;
 use crate::prelude::*;
 use bevy::platform::collections::HashMap;
 use bevy::prelude::*;
-use rand::{Rng, SeedableRng, rngs::SmallRng};
 use std::any::{Any, TypeId};
 use std::fmt::Debug;
 
@@ -32,7 +31,7 @@ impl DialogueRunnerBuilder {
                 yarn_project,
             )),
             asset_providers: HashMap::default(),
-            library: create_extended_standard_library(),
+            library: YarnLibrary::standard_library(),
             commands: YarnCommands::builtin_commands(commands),
             compilation: yarn_project.compilation().clone(),
             localizations: yarn_project.localizations().cloned(),
@@ -117,85 +116,5 @@ impl DialogueRunnerBuilder {
         }
 
         Ok(dialogue_runner)
-    }
-}
-
-fn create_extended_standard_library() -> YarnLibrary {
-    let mut library = YarnLibrary::standard_library();
-    library
-        .add_function("random", || SmallRng::from_os_rng().random_range(0.0..1.0))
-        .add_function("random_range", |min: f32, max: f32| {
-            if let Some(min) = min.as_int()
-                && let Some(max_inclusive) = max.as_int()
-            {
-                return SmallRng::from_os_rng().random_range(min..=max_inclusive) as f32;
-            }
-            SmallRng::from_os_rng().random_range(min..max)
-        })
-        .add_function("dice", |sides: u32| {
-            if sides == 0 {
-                return 1;
-            }
-            SmallRng::from_os_rng().random_range(1..=sides)
-        })
-        .add_function("round", |num: f32| num.round() as i32)
-        .add_function("round_places", |num: f32, places: u32| {
-            num.round_places(places)
-        })
-        .add_function("floor", |num: f32| num.floor() as i32)
-        .add_function("ceil", |num: f32| num.ceil() as i32)
-        .add_function("inc", |num: f32| {
-            if let Some(num) = num.as_int() {
-                num + 1
-            } else {
-                num.ceil() as i32
-            }
-        })
-        .add_function("dec", |num: f32| {
-            if let Some(num) = num.as_int() {
-                num - 1
-            } else {
-                num.floor() as i32
-            }
-        })
-        .add_function("decimal", |num: f32| num.fract())
-        .add_function("int", |num: f32| num.trunc() as i32);
-    library
-}
-
-trait FloatExt: Copy {
-    fn as_int(self) -> Option<i32>;
-    fn round_places(self, places: u32) -> Self;
-}
-
-impl FloatExt for f32 {
-    fn as_int(self) -> Option<i32> {
-        (self.fract() <= f32::EPSILON).then_some(self as i32)
-    }
-
-    fn round_places(self, places: u32) -> Self {
-        let factor = 10_u32.pow(places) as f32;
-        (self * factor).round() / factor
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn rounds_places() {
-        for (num, places, expected) in [
-            (1.0, 0, 1.0),
-            (1.2, 1, 1.2),
-            (0.4, 0, 0.0),
-            (43.132, 0, 43.0),
-            (1.1, 2, 1.1),
-            (123.123, 3, 123.123),
-            (-10.3, 1, -10.3),
-            (-11.99, 1, -12.0),
-        ] {
-            assert_eq!(expected, num.round_places(places));
-        }
     }
 }
