+++ crates/core/Cargo.toml
@@ -9,6 +9,9 @@ authors = [ "Jan Hohenheim <jan@hohenheim.ch>" ]
 license = "MIT OR Apache-2.0"
 description = "Core concepts for Yarn Spinner for Rust, the friendly tool for writing game dialogue"
 
+[target.wasm32-unknown-unknown.dependencies]
+getrandom = { version = "0.3", features = [ "wasm_js" ] }
+
 [features]
 default = [ "std" ]
 std = [ "prost/std" ]
@@ -16,7 +19,7 @@ serde = [ "dep:serde", "bevy?/serialize", "hashbrown/serde" ]
 bevy = [ "dep:bevy" ]
 
 [dependencies]
-rand = { version = "0.10", default-features = false, features = [ "thread_rng" ] }
+rand = { version = "0.9", default-features = false, features = [ "small_rng" ] }
 prost = { version = "0.12", default-features = false, features = [ "prost-derive" ] }
 serde = { version = "1", features = [ "derive" ], optional = true }
 bevy = { version = "0.18", default-features = false, optional = true, features = [ "std" ] }
