+++ examples/yarnspinner_without_bevy/assets/dialogue/custom_function.yarn
@@ -1,6 +1,8 @@
 title: CustomFunction
 ---
-Rust functions can be registered into your dialogue's library to make them runnable via Yarn.
-Let's take a look at a custom function now!
+The following are some outputs of built-in functions.
+random() = {random()}
+random_range(1,10) = {random_range(1,10)}
+There are more, but let's take a look at a custom function now.
 pow(2,3) = {pow(2,3)}
 ===
